20249257006/langchain/requirements.txt
SQLAlchemy>=1.4,<3
langchain-core>=0.2.43,<0.3
langsmith>=0.1.17,<0.2
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langchain-ollama>=0.1,<0.2
langchain-text-splitters>=0.2,<0.3
langchain-chroma==0.1.2
langchain-huggingface==0.0.3
chromadb>=0.5,<0.6
posthog>=2.4.0,<3.0.0
# chromadb 0.5.x tokenizers üst sınırı koyar; sentence-transformers ile çakışmayı önlemek için:
tokenizers==0.20.3
transformers>=4.41.0,<4.46.0
pypdf>=4,<5
python-dotenv>=1,<2
streamlit>=1.33,<2
sentence-transformers>=2.6,<3
torch>=2.2,<3
